# chore: Parallel clones for large data structures