Further work from https://github.com/lurk-lab/lurk-rs/pull/777.

This continues the story. pq observes that:
> I now believe, based on informal observation, that the loading in mmap-fix is slower than generating fresh params.

The long loading times were due to two separate issues. The first is a bug (fixed in https://github.com/lurk-lab/lurk-rs/pull/777), and the second is the lack of parallelization. This PR addresses the low hanging fruit by implementing parallel clones for the large data structures. We see ~2x speedup.

Before this PR:
```
bang!                                         24.328472s ├─────────────────────────────────────────────────────────────┤
  Instance::new                                535.148ms │
    <MultiFrame as StepCircuit>::synthesize     230.39ms  │
  read_abomonated                             15.416185s  ├──────────────────────────────────────┤
    decode                                       4.137ms  ┆
    clone                                     15.411941s  ├──────────────────────────────────────┤

```

After this PR:
```
bang!                                         18.039595s ├─────────────────────────────────────────────────────────────┤
  Instance::new                                530.639ms ├┤
    <MultiFrame as StepCircuit>::synthesize    233.376ms  │
  read_abomonated                              8.473104s   ├────────────────────────────┤
    decode                                         2.8ms   ┆
    clone                                      8.469151s   ├────────────────────────────┤

```